moved stock analyzer into old position
